requirements.txt
faster-whisper
Flask
ffmpeg-python
pydub
soundfile 